Fix a signed/unsigned warning

+++ src/wallet/wallet.cpp
@@ -1061,7 +1061,7 @@ void CWalletTx::GetAmounts(list<COutputEntry>& listReceived,
     }
 
     // Sent/received.
-    for (int i = 0; i < vout.size(); ++i)
+    for (unsigned int i = 0; i < vout.size(); ++i)
     {
         const CTxOut& txout = vout[i];
         // Skip special stake out
@@ -1089,7 +1089,8 @@ void CWalletTx::GetAmounts(list<COutputEntry>& listReceived,
                      this->GetHash().ToString());
             address = CNoDestination();
         }
-        COutputEntry output = {address, txout.nValue, i};
+
+        COutputEntry output = {address, txout.nValue, (int)i};
 
         // If we are debited by the transaction, add the output as a "sent" entry
         if (nDebit > 0)
